update to school portal

+++ database/migrations/2022_08_10_155124_create_results_table.php
@@ -15,10 +15,10 @@ class CreateResultsTable extends Migration
     {
         Schema::create('results', function (Blueprint $table) {
             $table->uuid('uuid')->primary()->unique();
-            $table->integer('ca1')->nullable();
-            $table->integer('ca2')->nullable();
-            $table->integer('ca3')->nullable();
-            $table->integer('exam')->nullable();
+            $table->double('ca1')->nullable();
+            $table->double('ca2')->nullable();
+            $table->double('ca3')->nullable();
+            $table->double('exam')->nullable();
             $table->foreignId('period_id')->constrained('periods')->onDelete('cascade');
             $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
             $table->foreignId('grade_id')->constrained('grades')->onDelete('cascade');
