+++ database/migrations/2023_01_05_080211_create_mid_terms_table.php
@@ -15,11 +15,11 @@ class CreateMidTermsTable extends Migration
     {
         Schema::create('mid_terms', function (Blueprint $table) {
             $table->uuid('uuid')->primary()->unique();
-            $table->integer('entry_1')->nullable();
-            $table->integer('entry_2')->nullable();
-            $table->integer('first_test')->nullable();
-            $table->integer('ca')->nullable();
-            $table->integer('project')->nullable();
+            $table->double('entry_1')->nullable();
+            $table->double('entry_2')->nullable();
+            $table->double('first_test')->nullable();
+            $table->double('ca')->nullable();
+            $table->double('project')->nullable();
             $table->foreignId('period_id')->constrained('periods')->onDelete('cascade');
             $table->foreignId('term_id')->constrained('terms')->onDelete('cascade');
             $table->foreignId('grade_id')->constrained('grades')->onDelete('cascade');
